backend/requirements.txt
alembic==1.13.1
autopep8==2.1.0
bcrypt==4.1.2
bidict==0.23.1
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-PyMongo==2.3.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
h11==0.14.0
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
mailjet-rest==1.3.4
Mako==1.3.3
MarkupSafe==2.1.5
numpy==1.26.4
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
py-crypto==0.0.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycrypto==2.6.1
PyJWT==2.8.0
pymongo==4.7.0
pyparsing==3.1.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
requests==2.31.0
rsa==4.9
simple-websocket==1.0.0
SQLAlchemy==2.0.29
tomli==2.0.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.2
wsproto==1.2.0
WTForms==3.1.2
